requirements.txt
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
kafka-python==2.0.2